refactor tui
